Split up Auth Tests
